book transfer refactored into service object

+++ app/controllers/user_destructions_controller.rb
@@ -21,14 +21,21 @@ class UserDestructionsController < ApplicationController
       # transfer books to another owner
       case params[:book_selection]
       when 'transfer'
-        transfer_books(owner, Owner.find_by(email: params[:transfer_to_owner]))
-        action_message = "books transferred to owner #{params[:transfer_to_owner]}"
+        # transfer_books(owner, Owner.find_by(email: params[:transfer_to_owner]))
+        # action_message = "books transferred to owner #{params[:transfer_to_owner]}"
+        result = BookTransfer.new(
+          owner,
+          Owner.find_by(email: params[:transfer_to_owner])
+        ).transfer
+        if result.transferred?
+          owner.destroy
+          redirect_to root_path, info: "Owner Removed; #{result.message}"
+        end
       when 'delete'
-        delete_books(owner)
-        action_message = 'books removed'
+        Book.destroy_by(owner:)
+        owner.destroy
+        redirect_to root_path, info: 'Owner Removed; books removed'
       end
-      owner.destroy
-      redirect_to root_path, info: "Owner Removed; #{action_message}"
     else
       redirect_to root_path
     end
@@ -36,20 +43,19 @@ class UserDestructionsController < ApplicationController
   # rubocop:enable Metrics/AbcSize
   # rubocop:enable Metrics/MethodLength
 
-  private
 
   # transfer books to another owner
   # this method smells of :reek:UtilityFunction
-  def transfer_books(old_owner, new_owner)
-    old_owner.books.each do |book|
-      book.update(owner_id: new_owner.id)
-    end
-  end
+  # def transfer_books(old_owner, new_owner)
+  #   old_owner.books.each do |book|
+  #     book.update(owner_id: new_owner.id)
+  #   end
+  # end
 
   # delete books
   # This method smells of :reek:UtilityFunction
-  def delete_books(existing_owner)
-    Book.destroy_by(owner: existing_owner)
-  end
+  # def delete_books(existing_owner)
+  #   Book.destroy_by(owner: existing_owner)
+  # end
 
 end
